Add HW src/dst tracking to load/store callbacks
This adds (attempted) tracking of the hardware source or
destination of loads/stores (for x86) to the pre/post load/store
callbacks.

Additionally, reorganizes a lot of the load/store callback code
to hopefully make it a bit easier to modify & maintain. This is
primarily useful because dodging argument limitations means there's
a _lot_ of effectively auto-generated code that exists to marshal
arguments.