First version of auto_scissors.py
You provide a replay and the name of a program that ran during that
replay and this script will create a scissors-ed version of the replay
just containing the part corresponding to the execution of that
program. Optionally, you can tell it the offset of the start of main
and the name of the module containing main and it will start the
scissors there instead.